requirements.txt
asgiref==3.5.2
Django==4.0.5
psycopg2-binary==2.9.3
sqlparse==0.4.2
djangorestframework
django-filter
django-versatileimagefield
python-magic-debian-bin==0.4.25
djangorestframework-simplejwt[crypto]
django-filter
